web/requirements.txt
# Global
Django==2.0.1
psycopg2

# Django core
django-split-settings==0.3.0

# App specific
django-polymorphic
django-markdownx

# Markdown
python-markdown-math
